エフェクト出す時間初期値変更

+++ Surprise_Party/Surprise_Party/SourceCode/SpriteEffect/SleepEffect/CSleepEffect.h
@@ -25,7 +25,7 @@ public:
 	const int			LINE_MAX			= 3;								//横の行最大数.
 	const float			SLEEP_Z_NUM			= 1;								//眠りマークの番号.
 	const float			ANGLE_MOVE_SPEED	= 5.0f;								//角度移動速度.
-	const float			APPEAR_TIME			= 30;								//表示時間.
+	const int			APPEAR_TIME			= 30;								//表示時間.
 
 	const float			ALPHA_SPEED			= 0.005f;							//透過速度.
 	const float			SCALE_SPEED			= 0.005f;							//拡縮速度.
